test(docs): pptr refactor
