+++ docs/scripts/util/index.ts
@@ -1,2 +1,6 @@
+export { checkLink } from './checkLink';
+export { crawlAllLinksFromAllPages } from './crawlAllLinksFromAllPages';
 export { getAllTypesData } from './getAllTypesData';
 export { getCatalog } from './getCatalog';
+export { pause } from './pause';
+export { runArrayPromiseInOrder } from './runArrayPromiseInOrder';
