+++ docs/scripts/util/pause.ts
@@ -1,0 +1,8 @@
+export async function pause(timeout: number) {
+  return new Promise((resolve) => {
+    console.log(`Pausing for ${timeout}`);
+    setTimeout(() => {
+      resolve(timeout);
+    }, timeout);
+  });
+}
